requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

langchain
langchain-community
langchain-text-splitters
langchain-huggingface

# pytz

sentence-transformers
faiss-cpu

huggingface-hub
